Prevent writing a file through a symlink
Because `fs.realpath` requires an actual file to exist, we can't rely on
it to detect if a symlink might point outside the output directory.

Traversing chains of symlinks, where the targets might be relative or
absolute, missing or not, gets complicated very quickly. And because
the behavior is generally surprising, of file creation writing through
to a pre-existing symlink's target, it seems preferable to disallow that
behavior.